pyproject.toml
[project]
name = "chat-with-docs"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Erick Cruz", email = "crzerick6@gmail.com"},
]
dependencies = ["langchain-openai>=0.3.16", "langchain-core>=0.3.59", "langchain-chroma>=0.2.3", "langchain-community>=0.3.23", "langgraph>=0.4.2", "python-dotenv>=1.1.0", "pypdf>=5.4.0", "fastapi>=0.115.12"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
